Implemented ResetPassword Flow

+++ database/migrations/2022_12_07_093423_create_verifiers_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('verifiers', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->tinyInteger('otp');
+            $table->integer('otp');
             $table->date('expiry')->nullable();
             $table->timestamps();
         });
